Fixes two major issues: now a different tx can be given for learning mode, and role information isn't ignored
We want to be able to perform different tasks on different keyspaces. That is, we want to train on keyspace A, evaluate independently on keyspace B, and predict on keyspace C. To permit this, these 3 functions must accept a Grakn transaction, since a transaction is bounded to a single option. This gives the option to to as described.

Due to a mistake caused by deleting members for the mutable _all_feature_types dictionary, role_type and role_direction were being ignored for all depths. As a consequence of fixing this, weights arrays for Combine ops need to be specified per-depth. This is because the starting concepts are provided without role inforation, and their features are directly concatenated with the aggregation output of their neighbours.