Significant changes for debugging
 Adds plenty of TensorFlow summary information, mostly histograms, for debugging purposes. This highlighted that values were exploding, which was due to un-normalised unixtime dates. After that it was clear that weights were not changing in aggregation or combine steps, therefore learning wasn't learning. Also it became apparent that most of the data was zeros.

One idea/benefit was rule out the impact of the features that are not needed for the toy example. This means that in general if a whole feature of the graph shouldn't be used, then that can be specified. This yields a much more compact way of declaring the processing elements for each feature type. However, this still didn't fix the issue with the lack of learning. We see that the loss plot shows essentially only noise around a baseline.