Grakn 1.5.3 migration (#61)
## What is the goal of this PR?

We have migrated kglib to use Grakn 1.5.3, rather than using the Grakn 1.5.0 snapshot used for the kglib prereleases. This has meant also upgrading kglib's dependency on grakn-client to version 1.5.2. This work cleans the dependencies significantly, which is especially helpful for the robustness of the CI pipeline and reliability for end-users.

## What are the changes implemented in this PR?

kglib now uses the new Graql syntax, including enforced commas, and the change from `relationship` to `relation`.

The animal trade dataset has been loaded into a Grakn 1.5.3 distribution for Mac, and is now used in the CI pipeline to fulfil the end-to-end test.

The interface of grakn-client has changed since the previous (unreleased) version that was previously being used, thus the code reflects these usage changes.

The README for KGCN and for the animal trade KGCN example have both been updated to reflect the Grakn version change.