Test deployment (#69)
## What is the goal of this PR?

We enable the proper testing of pip deployments. Previously this was using bazel, and therefore could access the dependencies declared there. This was not testing the dependencies assembled for pip. While this has been the case, it is likely that anyone trying to use kglib encountered errors importing `grakn.client`, unless they already had `grakn-client` installed in their python environment.

## What are the changes implemented in this PR?

- CI now directly calls the end-to-end test, testing against a version of `grakn-kglib` deployed to `repo.grakn.ai`, without using bazel whatsoever
- We saw that the end-to-end test did indeed fail due to the dependencies supplied to `assemble-pip` via `install_requires` being outdated. These dependencies have now been updated.
- The BUILD file for the end-to-end test became redundant and was deleted
- `http_file` access to `animaltrade_dist` is no longer needed within Bazel, and has been removed