KGCN II - Relation prediction using DeepMind's Graph Nets (#85)
## What is the goal of this PR?

Create a more versatile approach to KGCNs, one that can be customised to perform:
- Relation prediction
- Attribute value prediction
- Subgraph prediction
- Graph optimisation
- Generation of Concept embeddings

## What are the changes implemented in this PR?

Overwrites the entirety of the KGCN project with a more advanced version. This new approach uses a learned message-passing algorithm, implemented in TensorFlow. The algorithm can predict node, edge and global properties for any input graph.

Implemented is a method to predict new Relations in extracted subgraphs, picking from a provided set of candidate relations.

Solves #51, #78, #42