Change around end-to-end and deployment tests, add accuracy assertion (#90)
## What is the goal of this PR?

Ensure that the scope and intention of the CI jobs done on PRs vs the CI jobs done after merge make sense. All tests, including end-to-end, are needed before PRs are merged, which presently means in the `test` job. Jobs to ensure the deployment via pip works correctly are only executed on `master` after PRs are merged, and should be less extensive than the end-to-end test, but still check everything should run for a user.

## What are the changes implemented in this PR?

- Use an end-to-end test during the test job, and a much shorter end-to-end (to check all components work together) for the deployment test
- Test the accuracy of the learner, with a very low threshold