Test the health of links in markdown files (#97)
## What is the goal of this PR?

Avoid having to click through all of the links in Markdown files to manually test if they are broken. Instead, test the health of links in markdown files automatically.

## What are the changes implemented in this PR?

Adds a CI job using `awesome_bot` to report on the link health of all repo markdown files. This should fail if any links do not validate properly.